refactor: Move test processor package path according to main source
