+++ source-connector/src/main/java/sourceconnector/domain/log/factory/JSONLogFactory.java
@@ -1,4 +1,4 @@
-package sourceconnector.domain.factory;
+package sourceconnector.domain.log.factory;
 
 import sourceconnector.domain.log.JSONLog;
 import sourceconnector.domain.log.Log;
