+++ source-connector/src/main/java/sourceconnector/domain/log/factory/LogFactory.java
@@ -1,4 +1,4 @@
-package sourceconnector.domain.factory;
+package sourceconnector.domain.log.factory;
 
 import sourceconnector.domain.log.Log;
 import sourceconnector.domain.log.LogMetadata;
