+++ source-connector/src/test/java/sourceconnector/domain/processor/impl/TrimMapperProcessorTest.java
@@ -1,8 +1,8 @@
-package sourceconnector.service.processor.impl;
+package sourceconnector.domain.processor.impl;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import sourceconnector.domain.factory.JSONLogFactory;
+import sourceconnector.domain.log.factory.JSONLogFactory;
 import sourceconnector.domain.log.JSONLog;
 import sourceconnector.domain.log.Log;
 import sourceconnector.domain.log.LogMetadata;
