+++ source-connector/src/test/java/sourceconnector/service/batcher/LogBatcherTest.java
@@ -3,10 +3,11 @@ package sourceconnector.service.batcher;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import sourceconnector.domain.batch.MessageBatch;
-import sourceconnector.domain.factory.JSONLogFactory;
+import sourceconnector.domain.log.factory.JSONLogFactory;
 import sourceconnector.domain.log.Log;
+import sourceconnector.domain.pipeline.factory.FileBaseLogPipelineBuilder;
+import sourceconnector.domain.pipeline.factory.PipelineBuilder;
 import sourceconnector.repository.file.LocalFileRepository;
-import sourceconnector.domain.pipeline.FileBaseLogPipeline;
 import sourceconnector.domain.pipeline.Pipeline;
 import sourceconnector.domain.processor.impl.EmptyFilterProcessor;
 import sourceconnector.domain.processor.impl.TrimMapperProcessor;
@@ -14,6 +15,7 @@ import sourceconnector.domain.processor.impl.TrimMapperProcessor;
 import java.io.File;
 import java.nio.file.Path;
 import java.util.Collections;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -24,12 +26,12 @@ class LogBatcherTest {
   void nextBatchAtEmptyFile() {
     // given
     File file = Path.of("src/test/resources/sample-data/empty.ndjson").toFile();
-
-    Pipeline<Log> pipeline = FileBaseLogPipeline.create(
+    PipelineBuilder<Log> pipelineBuilder = new FileBaseLogPipelineBuilder();
+    Pipeline<Log> pipeline = pipelineBuilder.create(
       new LocalFileRepository(),
       file.getPath(),
       new JSONLogFactory(),
-      new TrimMapperProcessor(new JSONLogFactory()), new EmptyFilterProcessor()
+      List.of(new TrimMapperProcessor(new JSONLogFactory()), new EmptyFilterProcessor())
     );
     Batchable<Log> batcher = new LogBatcher(pipeline, 100);
 
@@ -46,11 +48,12 @@ class LogBatcherTest {
   void nextBatchTest() {
     File file = Path.of("src/test/resources/sample-data/empty-included.ndjson").toFile();
 
-    Pipeline<Log> pipeline = FileBaseLogPipeline.create(
+    PipelineBuilder<Log> pipelineBuilder = new FileBaseLogPipelineBuilder();
+    Pipeline<Log> pipeline = pipelineBuilder.create(
       new LocalFileRepository(),
       file.getPath(),
       new JSONLogFactory(),
-      new TrimMapperProcessor(new JSONLogFactory()), new EmptyFilterProcessor()
+      List.of(new TrimMapperProcessor(new JSONLogFactory()), new EmptyFilterProcessor())
     );
     Batchable<Log> batcher = new LogBatcher(pipeline, 3);
